Add pyglet integration
