+++ imgui/integrations/cocos2d.py
@@ -2,37 +2,14 @@
 from __future__ import absolute_import
 
 import cocos
-from pyglet.window import mouse
-from pyglet.window import key
 
 import imgui
+from .pyglet import PygletMixin
 from .opengl import FixedPipelineRenderer
 
 
-class ImguiLayer(cocos.layer.Layer):
+class ImguiLayer(PygletMixin, cocos.layer.Layer):
     is_event_handler = True
-
-    REVERSE_KEY_MAP = {
-        key.TAB: imgui.KEY_TAB,
-        key.LEFT: imgui.KEY_LEFT_ARROW,
-        key.RIGHT: imgui.KEY_RIGHT_ARROW,
-        key.UP: imgui.KEY_UP_ARROW,
-        key.DOWN: imgui.KEY_DOWN_ARROW,
-        key.PAGEUP: imgui.KEY_PAGE_UP,
-        key.PAGEDOWN: imgui.KEY_PAGE_DOWN,
-        key.HOME: imgui.KEY_HOME,
-        key.END: imgui.KEY_END,
-        key.DELETE: imgui.KEY_DELETE,
-        key.BACKSPACE: imgui.KEY_BACKSPACE,
-        key.RETURN: imgui.KEY_ENTER,
-        key.ESCAPE: imgui.KEY_ESCAPE,
-        key.A: imgui.KEY_A,
-        key.C: imgui.KEY_C,
-        key.V: imgui.KEY_V,
-        key.X: imgui.KEY_X,
-        key.Y: imgui.KEY_Y,
-        key.Z: imgui.KEY_Z,
-    }
 
     def __init__(self):
         super(ImguiLayer, self).__init__()
@@ -42,76 +19,9 @@ class ImguiLayer(cocos.layer.Layer):
 
         self._map_keys()
 
-    def _map_keys(self):
-        key_map = self.io.key_map
-
-        # note: we cannot use default mechanism of mapping keys
-        #       because pyglet uses weird key translation scheme
-        for value in self.REVERSE_KEY_MAP.values():
-            key_map[value] = value
-
     def on_enter(self):
         super(ImguiLayer, self).on_enter()
 
         if self.renderer is None:
             self.io.display_size = cocos.director.director.get_window_size()
             self.renderer = FixedPipelineRenderer()
-
-    def on_mouse_motion(self, x, y, dx, dy):
-        self.io.mouse_pos = x, self.io.display_size.y - y
-
-    def on_key_press(self, key, mods):
-        if key in self.REVERSE_KEY_MAP:
-            self.io.keys_down[self.REVERSE_KEY_MAP[key]] = True
-
-    def on_key_release(self, key, mods):
-        if key in self.REVERSE_KEY_MAP:
-            self.io.keys_down[self.REVERSE_KEY_MAP[key]] = False
-
-    def on_text(self, text):
-        io = imgui.get_io()
-
-        for char in text:
-            io.add_input_character(ord(char))
-
-    def on_mouse_drag(self, x, y, dx, dy, button, modifiers):
-        self.io.mouse_pos = x, self.io.display_size.y - y
-
-        if button == mouse.LEFT:
-            self.io.mouse_down[0] = 1
-
-        if button == mouse.MIDDLE:
-            self.io.mouse_down[1] = 1
-
-        if button == mouse.RIGHT:
-            self.io.mouse_down[2] = 1
-
-    def on_mouse_press(self, x, y, button, modifiers):
-        self.io.mouse_pos = x, self.io.display_size.y - y
-
-        if button == mouse.LEFT:
-            self.io.mouse_down[0] = 1
-
-        if button == mouse.MIDDLE:
-            self.io.mouse_down[1] = 1
-
-        if button == mouse.RIGHT:
-            self.io.mouse_down[2] = 1
-
-    def on_mouse_release(self, x, y, button, modifiers):
-        self.io.mouse_pos = x, self.io.display_size.y - y
-
-        if button == mouse.LEFT:
-            self.io.mouse_down[0] = 0
-
-        if button == mouse.MIDDLE:
-            self.io.mouse_down[1] = 0
-
-        if button == mouse.RIGHT:
-            self.io.mouse_down[2] = 0
-
-    def on_mouse_scroll(self, x, y, mods, scroll):
-        self.io.mouse_wheel = scroll
-
-    def on_resize(self, width, height):
-        self.io.display_size = width, height
